+++ cmd/cli/actions.go
@@ -36,7 +36,7 @@ func (a abiWrapper) String() string {
 	result := ""
 	for _, action := range a.ABI.Actions {
 		result += fmt.Sprintf("---\n%s\n\n", action.Name)
-		typ, found := findTypeByName(a.ABI.Types, action.Name)
+		typ, found := a.ABI.FindTypeByName(action.Name)
 		if !found {
 			result += fmt.Sprintf("  Error: Type not found for action %s\n", action.Name)
 			continue
@@ -47,13 +47,13 @@ func (a abiWrapper) String() string {
 			}
 		}
 
-		output, found := findOutputByID(a.ABI.Outputs, action.ID)
+		output, found := a.ABI.FindOutputByID(action.ID)
 		if !found {
 			result += fmt.Sprintf("No outputs for %s with id %d\n", action.Name, action.ID)
 			continue
 		}
 
-		typ, found = findTypeByName(a.ABI.Types, output.Name)
+		typ, found = a.ABI.FindTypeByName(output.Name)
 		if !found {
 			result += fmt.Sprintf("  Error: Type not found for output %s\n", output.Name)
 			continue
@@ -65,24 +65,6 @@ func (a abiWrapper) String() string {
 
 	}
 	return result
-}
-
-func findOutputByID(outputs []abi.TypedStruct, id uint8) (abi.TypedStruct, bool) {
-	for _, output := range outputs {
-		if output.ID == id {
-			return output, true
-		}
-	}
-	return abi.TypedStruct{}, false
-}
-
-func findTypeByName(types []abi.Type, name string) (abi.Type, bool) {
-	for _, typ := range types {
-		if typ.Name == name {
-			return typ, true
-		}
-	}
-	return abi.Type{}, false
 }
 
 func init() {
